Improve sso agent
* Remove the behaviour of redirecting to index.jsp after SLO
* Deprecate OIDCSSOAgentFilter class since the class name is misleading
* Introduce new class OpenIdSSOAgentFilter for OpenID sso
